+++ js/animations.js
@@ -1,0 +1,3 @@
+// изменение масштаба изображения;
+// применение одного из заранее заготовленных эффектов;
+// выбор глубины эффекта с помощью ползунка;
