+++ js/filter.js
@@ -1,0 +1,2 @@
+// Фильтрация изображений от других пользователей
+// возможно тоже в другом разделе нужно, постепенно пойму 8)
