+++ js/inputs.js
@@ -1,0 +1,3 @@
+// добавление текстового комментария;
+// добавление хэштегов.
+// возможно нужно в модуль функшнс
