+++ js/scale-control.js
@@ -1,0 +1,2 @@
+// При нажатии на кнопки .scale__control--smaller и .scale__control--bigger должно изменяться значение поля .scale__control--value;
+// возсожно все должно быть в модуле анимейшнс
